+++ src/test/java/org/veritas/assessment/biz/mapper/QuestionCommentReadLogMapperTest.java
@@ -57,21 +57,21 @@ class QuestionCommentReadLogMapperTest {
     void testFind() {
         int userId = 1;
         int projectId = 2;
-        Map<Integer, QuestionCommentReadLog> map = mapper.findLog(userId, projectId);
+        Map<Long, QuestionCommentReadLog> map = mapper.findLog(userId, projectId);
         assertEquals(4, map.size());
         assertEquals(4, map.get(3).getLatestReadCommentId());
-        Map<Integer, QuestionCommentReadLog> map2 = mapper.findLog(userId, projectId);
+        Map<Long, QuestionCommentReadLog> map2 = mapper.findLog(userId, projectId);
         // use cache, so they are same
         assertEquals(map.hashCode(), map2.hashCode());
     }
 
     @Test
     void testInsertOrUpdate() {
-        QuestionCommentReadLog log1 = new QuestionCommentReadLog(1, 2, 3, 4);
+        QuestionCommentReadLog log1 = new QuestionCommentReadLog(1, 2, 3L, 4);
         int result = mapper.addOrUpdate(log1);
         assertEquals(1, result);
 
-        QuestionCommentReadLog log2 = new QuestionCommentReadLog(1, 2, 3, 5);
+        QuestionCommentReadLog log2 = new QuestionCommentReadLog(1, 2, 3L, 5);
         result = mapper.addOrUpdate(log2);
         assertEquals(1, result);
     }
